refactor: Outcome 테스트 이동 및 계산로직 Outcome으로 다시 이동
