+++ src/test/java/blackjack/domain/gametable/OutcomeTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.gametable;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 public class OutcomeTest {
-/*
+
     @Test
     @DisplayName("base:21, counterpart:22 경우 -> base 승")
     void result0() {
@@ -56,5 +56,4 @@ public class OutcomeTest {
     void result6() {
         assertThat(Outcome.getInstance(Score.of(23), Score.of(22))).isEqualTo(Outcome.DRAW);
     }
-*/
 }
